app/requirements.txt
fastapi
uvicorn
pydantic
pytest
python-dotenv
openai
google-generativeai
# anthropic
